Second commit - правки замечаний, добавлен метод tokenExtractor.

+++ src/test/java/CreateOrderTest.java
@@ -7,6 +7,7 @@ import io.qameta.allure.junit4.DisplayName;
 import io.restassured.response.ValidatableResponse;
 import org.junit.After;
 import org.junit.Before;
+import org.junit.BeforeClass;
 import org.junit.Test;
 import java.util.ArrayList;
 import static org.junit.Assert.assertEquals;
@@ -15,9 +16,9 @@ import static org.junit.Assert.assertTrue;
 @Epic("Работа с заказами")
 @Feature("Созадние заказа")
 public class CreateOrderTest {
-    UserClient userClient;
-    OrderClient orderClient;
-    IngredientClient ingredientClient;
+    static UserClient userClient;
+    static OrderClient orderClient;
+    static IngredientClient ingredientClient;
     User user;
     UserCredentials userCredentials;
     UserTokens userTokens;
@@ -25,29 +26,32 @@ public class CreateOrderTest {
     ArrayList ingredients;
     static Faker faker = new Faker();
 
+    @BeforeClass
+    public static void getReady() {
+        userClient = new UserClient();
+        ingredientClient = new IngredientClient();
+        orderClient = new OrderClient();
+    }
+
     @Before
     public void setUp() {
         user = UserGenerator.getRandom();
         userCredentials = new UserCredentials(user.getEmail(), user.getPassword());
-        userClient = new UserClient();
         userClient.createUser(user);
         ValidatableResponse loginTestUser = userClient.loginUser(userCredentials);
-        userTokens = new UserTokens(loginTestUser.extract().path("refreshToken"), loginTestUser.extract().path("accessToken"));
-        ingredientClient = new IngredientClient();
+        userTokens = userClient.tokensExtractor(loginTestUser);
         ValidatableResponse getIngredients = ingredientClient.getIngredients();
         ingredients = new ArrayList<>();
         ingredients.add(getIngredients.extract().path("data[0]._id"));
         ingredients.add(getIngredients.extract().path("data[1]._id"));
-        orderClient = new OrderClient();
         order = new Order(ingredients);
     }
 
     @After
     public void tearDown() {
-        userClient = new UserClient();
-        try {
+        if (userTokens.getAccessToken() != null) {
             userClient.deleteUser(userTokens.getAccessToken());
-        } catch (Exception exception) {
+        } else {
             System.out.println("Пользователя невозможно удалить, так как он не был создан.");
         }
     }
@@ -57,23 +61,20 @@ public class CreateOrderTest {
     @Description("Для ревьюера: В документации явно не указано, какой ответ и код ответа должна давать система при создании заказа без авторизации," +
             "поэтому проверка создана из предположения, что код ответа должен быть 401 и сообщение You should be authorised")
     public void orderCanNotBeCreatedWithoutAuth() {
-        orderClient = new OrderClient();
         ValidatableResponse createOrder = orderClient.createOrderWithoutAuth(order);
         assertEquals(401, createOrder.extract().statusCode());
         assertEquals("You should be authorised", createOrder.extract().path("message"));
     }
     @Test
-    @DisplayName("Создание заказа c авторизацией и ингридиентами")
+    @DisplayName("Создание заказа c авторизацией и ингредиентами")
     public void orderCanBeCreatedWithAuth() {
-        orderClient = new OrderClient();
         ValidatableResponse createOrder = orderClient.createOrderWithAuth(order, userTokens.getAccessToken());
         assertEquals(200, createOrder.extract().statusCode());
         assertTrue(createOrder.extract().path("name").toString().contains("бургер"));
     }
     @Test
-    @DisplayName("Создание заказа без ингридиентов")
+    @DisplayName("Создание заказа без ингредиентов")
     public void orderCanNotBeCreatedWithoutIngredients() {
-        orderClient = new OrderClient();
         ArrayList nullIngredients = new ArrayList<>();
         Order nullOrder = new Order(nullIngredients);
         ValidatableResponse createOrder = orderClient.createOrderWithAuth(nullOrder, userTokens.getAccessToken());
@@ -83,7 +84,6 @@ public class CreateOrderTest {
     @Test
     @DisplayName("Создание заказа c неверным хешем ингридиентов")
     public void orderCanNotBeCreatedWithWrongIngredients() {
-        orderClient = new OrderClient();
         ArrayList wrongIngredients = new ArrayList<>();
         wrongIngredients.add(faker.hashCode());
         Order wrongOrder = new Order(wrongIngredients);
