+++ src/test/java/GetUserOrdersTest.java
@@ -6,6 +6,7 @@ import io.qameta.allure.junit4.DisplayName;
 import io.restassured.response.ValidatableResponse;
 import org.junit.After;
 import org.junit.Before;
+import org.junit.BeforeClass;
 import org.junit.Test;
 import java.util.ArrayList;
 import static org.junit.Assert.assertEquals;
@@ -13,31 +14,36 @@ import static org.junit.Assert.assertEquals;
 @Epic("Работа с заказами")
 @Feature("Получение заказов конкретного пользователя")
 public class GetUserOrdersTest {
-    UserClient userClient;
-    OrderClient orderClient;
-    IngredientClient ingredientClient;
+    static UserClient userClient;
+    static OrderClient orderClient;
+    static IngredientClient ingredientClient;
     User user;
     UserCredentials userCredentials;
     UserTokens userTokens;
     Order order;
     ArrayList ingredients;
 
+    @BeforeClass
+    public static void getReady() {
+        userClient = new UserClient();
+        ingredientClient = new IngredientClient();
+        orderClient = new OrderClient();
+    }
+
     @Before
     public void setUp() {
         user = UserGenerator.getRandom();
         userCredentials = new UserCredentials(user.getEmail(), user.getPassword());
-        userClient = new UserClient();
         userClient.createUser(user);
         ValidatableResponse loginTestUser = userClient.loginUser(userCredentials);
-        userTokens = new UserTokens(loginTestUser.extract().path("refreshToken"), loginTestUser.extract().path("accessToken"));
+        userTokens = userClient.tokensExtractor(loginTestUser);
     }
 
     @After
     public void tearDown() {
-        userClient = new UserClient();
-        try {
+        if (userTokens.getAccessToken() != null) {
             userClient.deleteUser(userTokens.getAccessToken());
-        } catch (Exception exception) {
+        } else {
             System.out.println("Пользователя невозможно удалить, так как он не был создан.");
         }
     }
@@ -45,7 +51,6 @@ public class GetUserOrdersTest {
     @Test
     @DisplayName("Получение заказов конкретного авторизованного пользователя без сделанных заказов")
     public void authorizedUserWithoutCreatedOrderOrdersCanBeGot() {
-        orderClient = new OrderClient();
         ValidatableResponse getOrders = orderClient.getOrdersOfUserWithAuth(userTokens.getAccessToken());
         assertEquals(200, getOrders.extract().statusCode());
         assertEquals(true, getOrders.extract().path("success"));
@@ -55,8 +60,6 @@ public class GetUserOrdersTest {
     @DisplayName("Получение заказов конкретного авторизованного пользователя, который сделал 1 заказ")
     @Description("Дополнительная проверка того, что заказы передаются корректно")
     public void authorizedUserWithCreatedOrderOrdersCanBeGot() {
-        orderClient = new OrderClient();
-        ingredientClient = new IngredientClient();
         ValidatableResponse getIngredients = ingredientClient.getIngredients();
         ingredients = new ArrayList<>();
         ingredients.add(getIngredients.extract().path("data[0]._id"));
@@ -72,7 +75,6 @@ public class GetUserOrdersTest {
     @Test
     @DisplayName("Получение заказов конкретного пользователя без авторизации")
     public void unauthorizedUserOrdersCanNotBeGot() {
-        orderClient = new OrderClient();
         ValidatableResponse getOrders = orderClient.getOrdersOfUserWithoutAuth();
         assertEquals(401, getOrders.extract().statusCode());
         assertEquals(false, getOrders.extract().path("success"));
