+++ src/test/java/UpdateUserTest.java
@@ -5,41 +5,44 @@ import io.qameta.allure.junit4.DisplayName;
 import io.restassured.response.ValidatableResponse;
 import org.junit.After;
 import org.junit.Before;
+import org.junit.BeforeClass;
 import org.junit.Test;
 import static org.junit.Assert.assertEquals;
 
 @Epic("Работа с пользователями")
 @Feature("Изменение данных пользователя")
 public class UpdateUserTest {
-    UserClient userClient;
+    static UserClient userClient;
     User user;
     UserCredentials userCredentials;
     UserTokens userTokens;
 
+    @BeforeClass
+    public static void getReady() {
+        userClient = new UserClient();
+    }
+
     @Before
     public void setUp() {
-        userClient = new UserClient();
         user = UserGenerator.getRandom();
         userCredentials = new UserCredentials(user.getEmail(), user.getPassword());
         userClient.createUser(user);
         ValidatableResponse loginTestUser = userClient.loginUser(userCredentials);
-        userTokens = new UserTokens(loginTestUser.extract().path("refreshToken"), loginTestUser.extract().path("accessToken"));
+        userTokens = userClient.tokensExtractor(loginTestUser);
     }
 
     @After
     public void tearDown() {
-        userClient = new UserClient();
-        try {
+        if (userTokens.getAccessToken() != null) {
             userClient.deleteUser(userTokens.getAccessToken());
-        } catch (Exception exception) {
+        } else {
             System.out.println("Пользователя невозможно удалить, так как он не был создан.");
         }
     }
 
     @Test
-    @DisplayName("Изменение пароля без авторизациии")
+    @DisplayName("Изменение пароля без авторизации")
     public void updateUserPasswordWithoutAuth() {
-        userClient = new UserClient();
         User userWithNewPassword = User.builder()
                 .password(user.getPassword() + "a")
                 .name(user.getName())
@@ -50,9 +53,8 @@ public class UpdateUserTest {
         assertEquals("You should be authorised", updateUser.extract().path("message"));
     }
     @Test
-    @DisplayName("Изменение имени пользователя без авторизациии")
+    @DisplayName("Изменение имени пользователя без авторизации")
     public void updateUserNameWithoutAuth() {
-        userClient = new UserClient();
         User userWithNewPassword = User.builder()
                 .password(user.getPassword())
                 .name(user.getName() + "a")
@@ -65,7 +67,6 @@ public class UpdateUserTest {
     @Test
     @DisplayName("Изменение email пользователя без авторизации")
     public void updateUserEmailWithoutAuth() {
-        userClient = new UserClient();
         User userWithNewPassword = User.builder()
                 .password(user.getPassword())
                 .name(user.getName())
@@ -78,7 +79,6 @@ public class UpdateUserTest {
     @Test
     @DisplayName("Изменение пароля с авторизацией")
     public void updateUserPasswordWithAuth() {
-        userClient = new UserClient();
         User userWithNewPassword = User.builder()
                 .password(user.getPassword() + "a")
                 .name(user.getName())
@@ -90,9 +90,8 @@ public class UpdateUserTest {
         assertEquals(userWithNewPassword.getName(), updateUser.extract().path("user.name"));
     }
     @Test
-    @DisplayName("Изменение имени пользователя c авторизациией")
+    @DisplayName("Изменение имени пользователя c авторизацией")
     public void updateUserNameWithAuth() {
-        userClient = new UserClient();
         User userWithNewName = User.builder()
                 .password(user.getPassword())
                 .name(user.getName() + "a")
@@ -106,7 +105,6 @@ public class UpdateUserTest {
     @Test
     @DisplayName("Изменение email пользователя с авторизацией")
     public void updateUserEmailWithAuth() {
-        userClient = new UserClient();
         User userWithNewEmail = User.builder()
                 .password(user.getPassword())
                 .name(user.getName())
